do no use compiler jinja to avoid unneeded hash
When {{ compiler('c') }} is used in a recipe, even when not used for a
particular platform, the compiler contributes to the hash. This is not needed
here and should be avoided. Explicitly specifying gcc_linux-ppc64le pulls in
the needed sysroot and run_exports while avoiding introducing a hash in the
package name.